% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{amax}
\alias{amax}
\alias{amin}
\title{Array searching}
\usage{
amax(a, axis = NULL)

amin(a, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{axis}{Axis or axes along which to operate. By default (\code{NULL}), flattened input is used.}
}
\value{
Maximum of \code{a}. If \code{axis} is \code{NULL}, the result is a scalar value. If \code{axis} is an int, the result is an array of dimension \code{\link{ndim}} - 1. If \code{axis} is a tuple, the result is an array of dimension \code{\link{ndim}} - \code{length(axis)}.

Minimum of \code{a}. If \code{axis} is \code{NULL}, the result is a scalar value. If \code{axis} is an int, the result is an array of dimension \code{\link{ndim}} - 1. If \code{axis} is a tuple, the result is an array of dimension \code{\link{ndim}} - \code{length(axis)}.
}
\description{
Return the maximum of an array or maximum along an axis.\cr
Return the minimum of an array or minimum along an axis.
}
\details{
This function corresponds to \code{amax()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.amax.html}{see}).

This function corresponds to \code{amin()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.amin.html}{see}).
}
\examples{
a <- marray(c(0:3), dim = c(2, 2))
amax(a)
amax(a, axis = 1)
amax(a, axis = 2)

amin(a)
amin(a, axis = 1)
amin(a, axis = 2)

}
