% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{argmax}
\alias{argmax}
\alias{argmin}
\title{Array searching}
\usage{
argmax(a, axis = NULL)

argmin(a, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{axis}{The axis along which to search for values. By default (\code{NULL}), the index is into the flattened array.}
}
\value{
Array of indices into the array. It has the same shape as \code{a} with the dimension along axis removed.
}
\description{
Get indices of maximum or minimum values.
}
\details{
This function corresponds to \code{argmax()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.argmax.html}{see}).

This function corresponds to \code{argmin()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.argmin.html}{see}).
}
\examples{
a <- marray(c(1:8, 8), dim = c(3, 3, 1))
argmax(a)
argmax(a, axis = 1)
argmax(a, axis = 2)
argmax(a, axis = 3)

}
