% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{argwhere}
\alias{argwhere}
\title{Array searching}
\usage{
argwhere(a, condition)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to keep.}
}
\value{
Indices of elements of \code{a} that meet \code{condition}, otherwise \code{NULL}.
}
\description{
Find the indices of those elements of an array that satisfy conditions.
}
\details{
This function corresponds partially to \code{argwhere()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.argwhere.html}{see}).
}
\examples{
a <- marray(0:5, dim = c(2, 3))
argwhere(a, a > 1)

}
