% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.r
\name{sparse_to_onehot}
\alias{sparse_to_onehot}
\alias{onehot_to_sparse}
\title{Masked Array}
\usage{
sparse_to_onehot(a)

onehot_to_sparse(a, axis = -1)
}
\arguments{
\item{a}{An array.}

\item{axis}{The axis along which to search for values.}
}
\value{
A binary array with each value aka label form \code{a} encoded on a separate position of the last axis.
}
\description{
Converts a sparse array to an one-hot encoded array.
}
\examples{
a <- random_int(dim = c(5, 5), min = 0, max = 3)
sparse_to_onehot(a)

}
\seealso{
\code{\link{argmax}}
}
