% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{array_split}
\alias{array_split}
\title{Array splitting}
\usage{
array_split(a, indices_or_sections, axis = 1L)
}
\arguments{
\item{a}{An array to be divided into sub-arrays.}

\item{indices_or_sections}{An integerish vector indicating the type of splitting.
A single integer, N, denotes that the array \code{a} will be divided into N arrays along \code{axis}.
A sorted vector of integers denotes the entries where along \code{axis} the array \code{a} is split.}

\item{axis}{The axis along which to split, default is \code{1}.}
}
\value{
A list of sub-arrays.
}
\description{
Split an array into multiple sub-arrays along an axis.
}
\details{
This function corresponds to \code{array_split()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.split.html}{see}).
}
\examples{
a <- marray(1:8)
array_split(a, indices_or_sections = 3)
array_split(a, indices_or_sections = c(3, 5, 6))

a <- marray(1:12, dim = c(4, 3))
array_split(a, indices_or_sections = 2)
array_split(a, indices_or_sections = 2, axis = 2L)

a <- marray(1:24, dim = c(4, 3, 2), order = "F")
array_split(a, indices_or_sections = 2)
array_split(a, indices_or_sections = 2, axis = 2L)

}
