% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{atleast_1d}
\alias{atleast_1d}
\title{Convert inputs to arrays with at least one dimension}
\usage{
atleast_1d(..., order = c("C", "F"))
}
\arguments{
\item{...}{Any number of objects that are coerced into at least 1-D arrays.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array, or list of arrays, each with at least one dimension.
}
\description{
Convert inputs to arrays with at least one dimension
}
\details{
This function corresponds to \code{atleast_1d()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.atleast_1d.html}{see}).
}
