% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{atleast_3d}
\alias{atleast_3d}
\title{Convert inputs to arrays with at least three dimensions}
\usage{
atleast_3d(..., order = c("C", "F"))
}
\arguments{
\item{...}{Any number of objects that are coerced into at least 3-D arrays.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array, or list of arrays, each with at least three dimensions.
}
\description{
Convert inputs to arrays with at least three dimensions
}
\details{
This function corresponds to \code{atleast_3d()} from NumPy.
}
