% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.r
\name{axesindices}
\alias{axesindices}
\title{Extract indices of axes of an array}
\usage{
axesindices(a, value)
}
\arguments{
\item{a}{An array.}

\item{value}{Any number of values to search for in \code{a}.}
}
\value{
A list of the index or indices of the dimension space of \code{value} within \code{a}, otherwise \code{NA}.
}
\description{
Extract indices of axes of an array
}
\details{
If there is no match between \code{value} and the values of \code{a}, \code{NA} will be returned.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
axesindices(a, value = c(11, 2, 23))

a <- marray(sample(6, 24, replace = TRUE), dim = c(4, 3, 2))
axesindices(a, value = c(4, 6, 0))

a <- ones(dim = c(4, 3, 2))
axesindices(a, value = 1)

}
