% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{bartlett}
\alias{bartlett}
\title{Array window}
\usage{
bartlett(M)
}
\arguments{
\item{M}{Number of points in the output window. If zero or less, an empty array is returned.}
}
\value{
The triangular window, with the maximum value normalized to one (the value one appears only if the number of samples is odd), with the first and last samples equal to zero.
}
\description{
Return the Bartlett window.
}
\details{
This function corresponds to \code{bartlett()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.bartlett.html}{see}).\cr
The Bartlett window is defined as

\deqn{w(n) = \frac{2}{M-1} \left(\frac{M-1}{2} - \left|n - \frac{M-1}{2}\right|\right)}
}
\references{
Bartlett M. S. (1950): Periodogram Analysis and Continuous Spectra, In: Biometrika 37(1-2), pp. 1-16.
}
\seealso{
\code{\link{blackman}}, \code{\link{hamming}}, \code{\link{hanning}}, \code{\link{kaiser}}
}
