% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{bitwise}
\alias{bitwise}
\title{Bitwise array operation}
\usage{
bitwise(..., op = c("and", "or", "xor"))
}
\arguments{
\item{...}{Any number of arrays all with identical shapes.}

\item{op}{An operator indicating the type of bitwise operation.
* \code{and}: Bitwise AND concatenation
* \code{or}: Bitwise OR concatenation
* \code{xor}: Bitwise XOR concatenation}
}
\value{
An array with the same shape as the given arrays with bitwise combined elements.
}
\description{
Bitwise AND, OR, or XOR concatenation of position equal elements of the given arrays.
}
\examples{
a <- marray(c(0, 0, 0, 1, 1, 1, 0, 0, 1), dim = c(3, 3, 1))
b <- marray(c(0, 1, 0, 0, 1, 1, 0, 1, 0), dim = c(3, 3, 1))
c <- marray(c(0, 0, 1, 0, 1, 0, 0, 1, 1), dim = c(3, 3, 1))
bitwise(a, b, c, op = "and")
bitwise(a, b, c, op = "or")
bitwise(a, b, c, op = "xor")

}
