% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{broadcastDIM}
\alias{broadcastDIM}
\title{Retrieve broadcast dimensions}
\usage{
broadcastDIM(..., axis = NULL)
}
\arguments{
\item{...}{Any number of objects.}

\item{axis}{The axis along operations like binding are planned.}
}
\value{
The dimension all objects must meet in order for the operation to be performed.
}
\description{
Retrieve broadcast dimensions
}
\seealso{
\code{\link{reshape_broadcast}}.
}
