% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.r
\name{clip}
\alias{clip}
\title{Array clip}
\usage{
clip(a, ..., a_min, a_max)
}
\arguments{
\item{a}{An array.}

\item{...}{Indexing instructions with or without letters in form of \code{name = value} pairs. The names of the arguments specify the dimensions and the values its values.}

\item{a_min}{Minimum value.}

\item{a_max}{Maximum value.}
}
\value{
The array \code{a}, but where values < \code{a_min} are replaced with \code{a_min}, and those > \code{a_max} with \code{a_max}.
}
\description{
Clip (limit) the values in an array.
}
\details{
This function corresponds partially to \code{clip()} from NumPy.
In opposite to NumPy function a slice of \code{a} can be specified to which limitation of the values refers.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
# Limit entire array
clip(a, a_min = 3, a_max = 7)
# Limit only a region of the array
clip(a, i = 1:2, j = 2:3, a_min = 3, a_max = 7)

}
