% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.r
\name{copyto}
\alias{copyto}
\title{Array copying}
\usage{
copyto(dst, src, dst_axis_index = NULL, src_axis_index = NULL)
}
\arguments{
\item{dst}{The array to which values are copied.}

\item{src}{The array from which values are copied.}

\item{dst_axis_index}{Axis indexing instructions for \code{dst}. Specifies the area in \code{dst} to which values are copied.}

\item{src_axis_index}{Axis indexing instructions for \code{src}. Specifies the area in \code{src} from which values are copied.}
}
\value{
The array \code{dst}.
}
\description{
Copy values from one array to another.
}
\details{
This function corresponds partially to \code{copyto()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.copyto.html}{see}).
}
\examples{
a <- zeros(dim = c(4, 3, 2))
b <- ones(dim = c(3, 3))
copyto(a, b, dst_axis_index = list(i = 2:3, j = 1:2), src_axis_index = list(i = 1:2, j = 2:3))

}
