% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{count_nonzero}
\alias{count_nonzero}
\title{Array counting}
\usage{
count_nonzero(a, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{axis}{The axis or tuple of axes along which to count non-zeros. By default (\code{NULL}), meaning that non-zeros will be counted along a flattened version of \code{a}.}
}
\value{
The number of non-zeros along a given axis. Otherwise, the total number of non-zeros in the array.
}
\description{
Counts the number of non-zero values in an array.
}
\details{
This function corresponds to \code{count_nonzero()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.count_nonzero.html}{see}).\cr
If axis is given, the result array has the same shape as \code{a} with the dimension along axis removed.
}
\examples{
a <- marray(c(0, 1, 7, 0, 3, 0, 2, 19), dim = c(2, 4, 1))
count_nonzero(a)
count_nonzero(a, axis = 1)
count_nonzero(a, axis = 2)

}
