% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{DIM}
\alias{DIM}
\title{Retrieve dimensions of an object or its length}
\usage{
DIM(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The number of dimensions or in cases of an atomic object the length.
}
\description{
Retrieve dimensions of an object or its length
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
