% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{dimC<-}
\alias{dimC<-}
\title{Dimensions of an object}
\usage{
dimC(x) <- value
}
\arguments{
\item{x}{An object to set dimensions on.}

\item{value}{An integerish vector of new dimensions.}
}
\value{
The (redimensioned) object \code{x}.
}
\description{
Set the dimensions of an object in row-major order.
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
