% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.r
\name{dsplit}
\alias{dsplit}
\title{Array splitting}
\usage{
dsplit(a, indices_or_sections)
}
\arguments{
\item{a}{An array to be divided into sub-arrays.}

\item{indices_or_sections}{An integerish vector indicating the type of splitting.
A single integer, N, denotes that the array \code{a} will be divided into N arrays along \code{axis}.
A sorted vector of integers denotes the entries where along \code{axis} the array \code{a} is split.}
}
\value{
A list of sub-arrays.
}
\description{
Split an array into multiple sub-arrays along the 3rd axis (depth-wise).
}
\details{
This function corresponds to \code{dsplit()} from NumPy.
  It's equivalent to \code{\link{array_split}} along the third axis.
}
