% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{ensuredim}
\alias{ensuredim}
\alias{dropdim}
\title{Array enforcing and unravel}
\usage{
ensuredim(x)

dropdim(x, order = c("C", "F"))
}
\arguments{
\item{x}{A vector or array.}

\item{order}{The order in which elements of \code{x} should be read during flattening if \code{x} is a higher-dimensional array.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array of at least one dimension.

A vector.
}
\description{
Enforce array or convert to vector.
}
\examples{
v <- setNames(1:11, do.call(paste0, list("i", c(1:11))))
v
class(v)
v <- ensuredim(v)
class(v)
dim(v)
dimnames(v)

v <- dropdim(v)
class(v)
names(v)
dim(v)
dimnames(v)

}
\references{
Implementation credits go partially to \url{https://github.com/cran/arrayhelpers}.
}
