% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{expand_dims}
\alias{expand_dims}
\title{Expand the shape of an array}
\usage{
expand_dims(a, axis = -1L)
}
\arguments{
\item{a}{An array.}

\item{axis}{Index position of the new axis or axes in the expanded array. Negative numbers count from the back.}
}
\value{
The expanded array \code{a} with new shape.
}
\description{
Insert new axis or axes that will appear at the axis positions in the expanded array shape.
}
\details{
This function corresponds to \code{expand_dims()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.expand_dims.html}{see}).
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
DIM(expand_dims(a))
DIM(expand_dims(a, axis = c(1, -2)))
DIM(expand_dims(a, axis = c(-2, 3, -1)))
DIM(expand_dims(a, axis = c(1, 1, 3, -1)))

}
\seealso{
\code{\link{squeeze}}.
}
