% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.r
\name{extract}
\alias{extract}
\title{Array searching}
\usage{
extract(a, condition)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to keep.}
}
\value{
An array with rank 1 containing elements of \code{a} that meet \code{condition}, otherwise invisible \code{NULL}.
}
\description{
Return elements of an array that satisfy conditions.
}
\details{
This function is an equivalent to \code{extract()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.extract.html}{see}).
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
values <- extract(a, condition = (a > 3) & (a <= 11))
values

# Forward the values to axesindices() to get the coordinates of these values
axesindices(a, value = values)

}
