% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{eye}
\alias{eye}
\title{Array creation}
\usage{
eye(n, m = NULL, dimnames = NULL)
}
\arguments{
\item{n}{The number of rows.}

\item{m}{The number of columns, default \code{NULL}.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of length \code{n}.}
}
\value{
An identity matrix with n rows and n or rather m columns.
}
\description{
Create 2D identity matrix.
}
\details{
This function corresponds to \code{eye()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.eye.html}{see}).
}
