% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{flatnonzero}
\alias{flatnonzero}
\title{Array counting}
\usage{
flatnonzero(a)
}
\arguments{
\item{a}{An array.}
}
\value{
Vector containing the indices of non-zero elements within \code{a}.
}
\description{
Return indices that are non-zero in a flatten array.
}
\details{
This function corresponds to \code{flatnonzero()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.flatnonzero.html}{see}).\cr
}
\examples{
a <- marray(seq(-2, 2))
flatnonzero(a)

# Use the indices of the non-zero elements as an index vector to extract these elements
flatten(a)[flatnonzero(a)]

}
