% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fmeta<-,MArray-method}
\alias{fmeta<-,MArray-method}
\title{Replacement method for the 'fmeta' slot of an MArray object}
\usage{
\S4method{fmeta}{MArray}(object) <- value
}
\arguments{
\item{object}{An object of class MArray}

\item{value}{An item you wish to assign to the 'fmeta' slot}
}
\value{
An MArray object with its 'fmeta' slot updated
}
\description{
Replacement method for the 'fmeta' slot of an MArray object
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma2 <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 309)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
fmeta(ma_obj) <- ma2$fmeta
}
