% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fmeta}
\alias{fmeta}
\alias{fmeta<-}
\title{Accessor function for the 'fmeta' slot of an MArray object}
\usage{
fmeta(object)

fmeta(object) <- value
}
\arguments{
\item{object}{An object of class MArray}

\item{value}{Replacement item}
}
\description{
Accessor function for the 'fmeta' slot of an MArray object

Replacement method for the 'fmeta' slot of an MArray object
}
