% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{\%null\%}
\alias{\%null\%}
\title{NULL assignment operator}
\usage{
x \%null\% y
}
\arguments{
\item{x}{object that is checked for \code{NULL}.}

\item{y}{result if \code{x} is \code{NULL}.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
NULL assignment operator
}
\details{
Checking for \code{NULL} is very common practice in R. This operator shortens the handling of such checks and the return of corresponding values.
}
