% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{hamming}
\alias{hamming}
\title{Array window}
\usage{
hamming(M)
}
\arguments{
\item{M}{Number of points in the output window. If zero or less, an empty array is returned.}
}
\value{
The window, with the maximum value normalized to one (the value one appears only if the number of samples is odd).
}
\description{
Return the Hamming window.
}
\details{
This function corresponds to \code{hamming()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.hamming.html}{see}).\cr
The Hamming window is defined as

\deqn{w(n) = 0.54 - 0.46\cos\left(\frac{2\pi{n}}{M-1}\right) \qquad 0 \leq n \leq M-1}
}
\references{
Blackman, R.B. and Tukey, J.W. (1958): The measurement of power spectra, Dover Publications, New York.
}
\seealso{
\code{\link{bartlett}}, \code{\link{blackman}}, \code{\link{hanning}}, \code{\link{kaiser}}
}
