% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{hstack}
\alias{hstack}
\title{Array stack}
\usage{
hstack(..., input_shape = NULL, order = c("C", "F"))
}
\arguments{
\item{...}{Any number of objects that are combined into an array. The objects can be packed into a \code{list}. The dimensions of these objects must be equal, excluding axis 2, if they are not to be coerced into a certain \code{input_shape}.}

\item{input_shape}{The dimension the input objects are to be coerced. By default \code{NULL}, the original dimensions are used.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array formed by stacking the given arrays.
}
\description{
Stack arrays in sequence horizontally (column-wise).
}
\details{
This function corresponds to \code{hstack()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.hstack.html}{see}).
  It's equivalent to \code{\link{mabind}} along the second axis.
}
