% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.r
\name{insert}
\alias{insert}
\title{Array insertion}
\usage{
insert(a, ..., axis = -1L, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{...}{Any number of objects inserted into or appended to \code{a}.}

\item{axis}{The axis along which to insert the objects.}

\item{order}{The order according to which the respective elements of the objects are read during insertion or appending.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The array \code{a} with objects inserted or appended.
}
\description{
Insert an object into an array.
}
\details{
This function corresponds partially to \code{insert()} and \code{append()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.insert.html}{see}).
}
\examples{
# original array
a <- array(seq.int(2 * 3 * 4), dim = c(2, 3, 4))
# slice to be added to the second axis
b <- array(100L + seq.int(2 * 1 * 4), dim = c(2, 1, 4))
insert(a, b, axis = 2L)

}
