% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.r
\name{kaiser}
\alias{kaiser}
\title{Array window}
\usage{
kaiser(M, beta)
}
\arguments{
\item{M}{Number of points in the output window. If zero or less, an empty array is returned.}

\item{beta}{Shape parameter for window. The Kaiser can approximate many other windows by varying the beta parameter:
\itemize{
\item \code{0}: Rectangular.
\item \code{5}: Similar to a Hamming.
\item \code{6}: Similar to a Hanning.
\item \code{8.6}: Similar to a Blackman.
}}
}
\value{
The window, with the maximum value normalized to one (the value one appears only if the number of samples is odd).
}
\description{
Return the Kaiser window.
}
\details{
This function corresponds to \code{kaiser()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.kaiser.html}{see}).\cr
The Kaiser window is defined as

\deqn{w(n) = I_0\left(\beta \sqrt{1-\frac{4n^2}{(M-1)^2}} \right)/I_0(\beta)} with \eqn{\quad -\frac{M-1}{2} \leq n \leq \frac{M-1}{2}}
}
\references{
Kaiser, J. F. (1966): Digital Filters. In: F.F. Kuo and J.F. Kaiser (Eds.), Systems analysis by digital computer, pp. 218-285. New York: John Wiley and Sons.
}
\seealso{
\code{\link{bartlett}}, \code{\link{blackman}}, \code{\link{hamming}}, \code{\link{hanning}}
}
