% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.r
\name{mabind}
\alias{mabind}
\title{Array binding}
\usage{
mabind(..., input_shape = NULL, axis = -1, order = c("C", "F"))
}
\arguments{
\item{...}{Any number of objects that are combined into an array. The objects can be packed into a \code{list}. The dimensions of these objects must be equal, excluding axis, if they are not to be coerced into a certain \code{input_shape}.}

\item{input_shape}{The dimension the input objects are to be coerced. By default \code{NULL}, the original dimensions are used.}

\item{axis}{The dimension along the objects are combined. By default (\code{-1}), the last dimension is used for binding the arrays.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array as a combination of all input arrays along a specified dimension.
}
\description{
Combine arrays along a specified dimension.
}
\details{
This function corresponds to \code{concatenate()} from NumPy.
}
\examples{
a <- marray(1:11)
b <- marray(2:23)
mabind(a, b)

a <- marray(1:12, dim = c(4, 3))
b <- marray(1:12, dim = c(4, 3, 1))
mabind(a, b)

a1 <- marray(1:24, dim = c(4, 3, 2), order = "F"); a1
a2 <- marray(-1:-24, dim = c(4, 3, 2), order = "F"); a2
a3 <- marray(sample(24), dim = c(4, 3, 2), order = "F"); a3
mabind(a1, a2, a3) # output is an 4x3x6 array
mabind(a1, a2, a3, axis = 1) # output is an 12x3x2 array
mabind(a1, a2, a3, axis = 2) # output is an 4x9x2 array
}
