% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{madiff}
\alias{madiff}
\title{Array math}
\usage{
madiff(a, n = 1, axis = -1, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{n}{The number of times values are differenced. If zero, the input is returned as-is.}

\item{axis}{The axis along which the difference is taken, default is the last axis.}

\item{order}{The order in which elements of data should be read during operation.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An array with the n-th differences. The shape of the output is the same as \code{a} except along \code{axis} where the dimension is smaller by \code{n}.
}
\description{
Calculate the n-th discrete difference along the given axis.
}
\details{
This function corresponds to \code{diff()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.diff.html}{see}).
}
\examples{
a <- marray(c(3, 4, 8, 2, 6, 5), dim = c(2, 3))
madiff(a, axis = 1)
madiff(a, axis = 2)

a <- marray(c(11, 2, 4, 10, 13, 5, 1, 8, 3, 9, 10, 11, 2, 19, 4, 4, 10, 11, 23, 67, 69, 47, 21, 23), dim = c(4, 3, 2))
madiff(a)
madiff(a, axis = 1)
madiff(a, axis = 2)

}
