% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{maidentity}
\alias{maidentity}
\title{Array creation}
\usage{
maidentity(n, dimnames = NULL)
}
\arguments{
\item{n}{Number of rows and columns in \emph{n x n} output.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of length \code{n}.}
}
\value{
\emph{n x n} array with its main diagonal set to one, and all other elements to zero.
}
\description{
Return the identity array, a square array with ones on the main diagonal.
}
\details{
This function corresponds to \code{identity()} from NumPy.
}
