% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMArrayDataSet.R
\name{MakeMArrayDataSet}
\alias{MakeMArrayDataSet}
\title{Simulate a Micro array data set with accompanying metadata}
\usage{
MakeMArrayDataSet(n_samples = 10L, n_features = 6L, with_seed = 8675309)
}
\arguments{
\item{n_samples}{The number of samples to simulate (must be a positive, even integer value that is no greater than '26')}

\item{n_features}{The number of genes to simulate (must be a positive, even integer value)}

\item{with_seed}{A numeric input to set the seed of the session (defaults to '8675309')}
}
\value{
A list of 1 matrix ('marray' - simulated metadata), and 2 data frames (1 for sample metadata -- 'pmeta' -- and another for gene metadata -- 'fmeta')
}
\description{
Simulate microarray data by randomly drawing values from a normal distribution centered at 10 with a standard deviation of 5
}
\examples{
MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
}
