% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.r
\name{mamatrix}
\alias{mamatrix}
\title{Array to matrix}
\usage{
mamatrix(a, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{order}{The order in which elements of data should be read during combination.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
A two-dimensional array with combined matrices of the first two dimensions of \code{x}.
}
\description{
Shrinks an array by combining the respective first two dimensions of the array by columns or rows.
}
