% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat.r
\name{marepeat}
\alias{marepeat}
\title{Array repetition}
\usage{
marepeat(a, repeats, axis = NULL)
}
\arguments{
\item{a}{The input array.}

\item{repeats}{An integerish vector or list of integerish vectors indicating the number of repetitions for each element along \code{axis}.}

\item{axis}{The axis or axes along which to repeat values. Per default, all axes are addressed.}
}
\value{
An output array which has the same shape as \code{a}, except along \code{axis}.
}
\description{
Repeat elements of an array.
}
\details{
This function corresponds partially to \code{repeat()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.repeat.html}{see}) but extends its functionality to all axes of \code{a}.
}
\examples{
a <- marray(seq_len(4*4*2), dim = c(4, 4, 2))
marepeat(a, repeats = 2, axis = 1)
marepeat(a, repeats = c(2, 2, 3, 1), axis = 1)
marepeat(a, repeats = c(2, 2, 1, 1), axis = c(1, 2))
marepeat(a, repeats = c(2, 2, 1, 1), axis = 1) |> marepeat(repeats = c(2, 2, 1, 1), axis = 2)
marepeat(a, repeats = list(i = c(2, 1, 2, 1), j = c(1, 2, 3, 2)), axis = c(1, 2))

}
\seealso{
\code{\link{tile}}.
}
