% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{mareplace}
\alias{mareplace}
\title{Array replacement}
\usage{
mareplace(a, ..., oldvalue, newvalue)
}
\arguments{
\item{a}{An array.}

\item{...}{Indexing instructions with or without letters in form of \code{name = value} pairs. The names of the arguments specify the dimensions and the values its values.}

\item{oldvalue}{A vector of values to search for.}

\item{newvalue}{A vector of values to replace \code{oldvalue}.
A single value is used as a replacement for all values in \code{oldvalue}. Otherwise, the length of \code{newvalue} must be equal to the length of \code{oldvalue}.
In that case a value in \code{newvalue} at a certain position is used to replace the value in \code{oldvalue} at the same position.}
}
\value{
An array with replaced values.
}
\description{
Replace array elements.
}
\examples{
a <- marray(c(rep(1, 4), 2, rep(3, 5), 21:34), dim = c(4, 3, 2), order = "F")
mareplace(a, oldvalue = c(1, 200, 3), newvalue = c(100, 2, 300))

a <- zeros(dim = c(5, 4))
mareplace(a, i = 2:4, j = 2:3, oldvalue = 0, newvalue = 2)

}
