% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{marray<-,MArray-method}
\alias{marray<-,MArray-method}
\title{Replacement method for the 'marray' slot of an MArray object}
\usage{
\S4method{marray}{MArray}(object) <- value
}
\arguments{
\item{object}{An object of class MArray}

\item{value}{An item you wish to assign to the 'marray' slot}
}
\value{
An MArray object with its 'marray' slot updated
}
\description{
Replacement method for the 'marray' slot of an MArray object
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
marray(ma_obj) <- ma$marray + 100
}
