% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{marray}
\alias{marray}
\title{Accessor function for the 'marray' slot of an MArray object}
\usage{
marray(object)
}
\arguments{
\item{object}{An object of class MArray}
}
\description{
Accessor function for the 'marray' slot of an MArray object
}
