% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{marray}
\alias{marray}
\alias{marray<-}
\alias{marray,MArray-method}
\alias{marray<-,MArray-method}
\title{Accessor function for the 'marray' slot of an MArray object}
\usage{
marray(object)

marray(object) <- value

\S4method{marray}{MArray}(object)

\S4method{marray}{MArray}(object) <- value
}
\arguments{
\item{object}{An object of class MArray}

\item{value}{An item you wish to assign to the 'marray' slot}
}
\value{
A matrix of Micro array data

An MArray object with its 'marray' slot updated
}
\description{
Accessor function for the 'marray' slot of an MArray object

Replacement method for the 'marray' slot of an MArray object

Accessor function for the 'marray' slot of an MArray object

Replacement method for the 'marray' slot of an MArray object
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
marray(ma_obj)
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
marray(ma_obj) <- ma$marray + 100
}
