% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.r
\name{memberof}
\alias{memberof}
\alias{mask_rectangle}
\alias{mask_circle}
\title{Masked Array}
\usage{
memberof(a, condition)

mask_rectangle(dim, ..., size)

mask_circle(dim, ..., radius)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to search for.}

\item{dim}{Shape of the masked array.}

\item{...}{Indexing instructions with or without letters in form of \code{name = value} pairs. The names of the arguments specify the axis and the values its positions.}

\item{size}{An integerish vector or matrix indicating the size of the window around the center coordinates given by \code{...}.}

\item{radius}{An integerish vector indicating the radius around the coordinates given by \code{...}. A single value will be used for every \code{axis}.}
}
\value{
A binary array with the shape of \code{a}. If an element in \code{a} meets \code{condition} the result value is one at the corresponding position, otherwise it is zero.

A binary array with shape \code{dim}. All elements within the window around each center coordinate have the value one, the others zero.

A binary array with shape \code{dim}. All elements within the radius around of each coordinate have the value one, the others zero.
}
\description{
Create a binary array based on meeting conditions.

Create a binary array in rectangular form based on center coordinates and window sizes.

Create a binary array in circular form based on coordinates and radiuses.
}
\details{
The indexing instructions must correspond to the number of axis within \code{dim}. For every axis an integerish vector must be specified which can be used as center coordinates.
The \code{size} argument is either a single value used for every axis, or a vector, or a matrix. If a vector or matrix is given, the number of elements respectively
the number of columns must be also identical with the number of axis.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
memberof(a, condition = a \%in\% c(1, 2, 15, 25, 11, 23))
memberof(a, condition = (a > 11) & (a <= 23))

m <- mask_rectangle(dim = c(30, 30, 10), i = c(11, 23), j = c(15, 20), k = c(4, 7), size = 3)
slice(m, i = c(9:13), j = c(13:17), k = c(2:6))

m <- mask_circle(dim = c(10, 10), i = 4, j = 4, radius = 3)
m

}
\seealso{
\code{\link{where}}.
}
