% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.r
\name{masubset}
\alias{masubset}
\title{Array subsetting}
\usage{
masubset(a, subset, drop = FALSE)
}
\arguments{
\item{a}{An array.}

\item{subset}{Logical expression indicating elements to keep.}

\item{drop}{Boolean passed on to \code{[} indexing operator.}
}
\value{
A vector containing the elements of \code{a} which meet conditions.
}
\description{
Array subsetting
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
values <- masubset(a, subset = (a > 3) & (a <= 11))
values

# Forward the values to axesindices() to get the coordinates of these values
axesindices(a, value = values)

}
