% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{memberof}
\alias{memberof}
\title{Array searching}
\usage{
memberof(a, condition)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to search for.}
}
\value{
A binary array with the shape of \code{a}. If an element in \code{a} meets \code{condition} the result value is one at the corresponding position, otherwise it is zero.
}
\description{
Create a binary array based on meeting conditions.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
memberof(a, condition = a \%in\% c(1, 2, 15, 25, 11, 23))
memberof(a, condition = (a > 11) & (a <= 23))

}
\seealso{
\code{\link{where}}.
}
