% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.r
\name{moveaxis}
\alias{moveaxis}
\title{Array axis moving}
\usage{
moveaxis(a, source, destination)
}
\arguments{
\item{a}{An array.}

\item{source}{An integerish vector indicating the original axes to move. These must be unique.}

\item{destination}{An integerish vector indicating the destination positions for each of the original axes. These must also be unique.}
}
\value{
The array \code{a} with moved axes.
}
\description{
Move axes of an array to new positions while the other axes remain in their original order.
}
\details{
This function corresponds to \code{moveaxis()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.moveaxis.html}{see}).
}
\examples{
a <- marray(1:(3*4*5), dim = c(3, 4, 5), order = "F")
DIM(moveaxis(a, source = 1, destination = 3))
DIM(moveaxis(a, source = 3, destination = 1))
DIM(moveaxis(a, source = c(1, 2), destination = c(3, 1)))

a <- marray(1:24, dim = c(4, 3, 2, 1))
DIM(moveaxis(a, source = c(1, 4), destination = c(3, 2)))

}
