% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{ndim}
\alias{ndim}
\title{Number of dimensions}
\usage{
ndim(x)
}
\arguments{
\item{x}{A multidimensional data structure like array, matrix or data.frame.}
}
\value{
Number of dimensions.
}
\description{
Number of dimensions
}
\details{
This function corresponds to \code{ndarray.ndim} from NumPy.
}
