% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{ndmin}
\alias{ndmin}
\title{Array shape}
\usage{
ndmin(a, n, axis = -1L)
}
\arguments{
\item{a}{An array.}

\item{n}{The desired number of dimensions.}

\item{axis}{Index position of the new axis in the expanded array. Negative numbers count from the back.}
}
\value{
The array \code{a} with at least \code{n} dimensions.
}
\description{
Ensure minimum number of dimensions.
}
