% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{op-in-dim}
\alias{op-in-dim}
\alias{\%[in]\%}
\title{Value matching with dimension preserving}
\usage{
x \%[in]\% table
}
\arguments{
\item{x}{an array.}

\item{table}{the values to be matched against.}
}
\value{
A logical array with the same shape as \code{x}, indicating if a match was located for each element of \code{x}; thus the values are \code{TRUE} or \code{FALSE}.
}
\description{
Value matching with dimension preserving
}
\details{
The base R operator \code{\%in\%} doesn't preserve the dimension of the object values are matched against.
}
\examples{
a <- marray(sample(4, 12, replace = TRUE), dim = c(4, 3, 1))
a \%[in]\% c(3)
}
