% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{NULL assignment operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{object that is checked for \code{NULL}.}

\item{y}{result if \code{x} is \code{NULL}.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
NULL assignment operator
}
\details{
Checking for \code{NULL} is very common practice in R. This operator shortens the handling of such checks and the return of corresponding values.
  It is adopted from package \code{rlang}, also used in package \code{listarrays}.
}
\examples{
11 \%||\% 23
NULL \%||\% 23
}
