% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{op-zero-length-default}
\alias{op-zero-length-default}
\alias{\%|0|\%}
\title{Zero length assignment operator}
\usage{
x \%|0|\% y
}
\arguments{
\item{x}{object that is checked for \code{length}.}

\item{y}{result if length of \code{x} is \code{0}.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
Zero length assignment operator
}
