% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.r
\name{pad}
\alias{pad}
\title{Array padding}
\usage{
pad(
  a,
  pad_width,
  mode = c("constant", "edge", "linear_ramp", "maximum", "mean", "median", "minimum",
    "reflect", "symmetric", "wrap", "empty"),
  ...
)
}
\arguments{
\item{a}{An array.}

\item{pad_width}{Number of values padded to the edges of each axis. ((before_1, after_1), ... (before_N, after_N)) unique pad widths for each axis. One integerish value is a shortcut for before = after for all axes.}

\item{mode}{The technique to be used for padding an array.
\itemize{
\item \code{constant}: Pads with a constant value.
\item \code{edge}: Pads with the edge values of array.
\item \code{linear_ramp}: Pads with the linear ramp between end_value and the array edge value.
\item \code{maximum}: Pads with the maximum value of all or part of the vector along each axis.
\item \code{mean}: Pads with the mean value of all or part of the vector along each axis.
\item \code{median}: Pads with the median value of all or part of the vector along each axis.
\item \code{minimum}: Pads with the minimum value of all or part of the vector along each axis.
\item \code{reflect}: Pads with the reflection of the vector mirrored on the first and last values of the vector along each axis.
\item \code{symmetric}: Pads with the reflection of the vector mirrored along the edge of the array.
\item \code{wrap}: Pads with the wrap of the vector along the axis. The first values are used to pad the end and the end values are used to pad the beginning.
\item \code{empty}: Pads with undefined (\code{NA}) values.
}}

\item{...}{Optional arguments corresponding to \code{mode}.}
}
\value{
Padded array of rank equal to \code{a} with shape increased according to \code{pad_width}.
}
\description{
Pad an array.
}
\details{
This function corresponds to \code{pad()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.pad.html}{see}).
}
\examples{
a <- marray(1:5)
pad(a, pad_width = list(c(2, 3)), constant_values = list(c(4, 6)))

}
