% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{pmeta,MArray-method}
\alias{pmeta,MArray-method}
\title{Accessor function for the 'pmeta' slot of an MArray object}
\usage{
\S4method{pmeta}{MArray}(object)
}
\arguments{
\item{object}{An object of class MArray}
}
\value{
A dataframe of sample meta data
}
\description{
Accessor function for the 'pmeta' slot of an MArray object
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
pmeta(ma_obj)
}
