% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pmeta}
\alias{pmeta}
\alias{pmeta<-}
\title{Accessor function for the 'pmeta' slot of an MArray object}
\usage{
pmeta(object)

pmeta(object) <- value
}
\description{
Accessor function for the 'pmeta' slot of an MArray object

Replacement method for the 'pmeta' slot of an MArray object
}
