% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.r
\name{random_int}
\alias{random_int}
\alias{random_int_like}
\alias{random_dbl}
\alias{random_dbl_like}
\alias{random_normal}
\title{Array creation}
\usage{
random_int(dim = NULL, min = 0, max = 1, dimnames = NULL, order = c("C", "F"))

random_int_like(a, min = 0, max = 1, dimnames = NULL, order = c("C", "F"))

random_dbl(dim = NULL, min = 0, max = 1, dimnames = NULL, order = c("C", "F"))

random_dbl_like(a, min = 0, max = 1, dimnames = NULL, order = c("C", "F"))

random_normal(dim = NULL, mean = 0, sd = 1, order = c("C", "F"))
}
\arguments{
\item{dim}{Shape of the new array.}

\item{min, max}{Lower and upper limits of the generated random values.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of the length given by \code{dim} for that dimension.}

\item{order}{Whether to store multidimensional array in C- or Fortran-contiguous (row- or column-wise) order.}

\item{a}{An array whose shape is taken for the result array.}
}
\value{
An array with random values with the given shape.
}
\description{
Create array filled with random integer or double values.
}
\details{
This function corresponds roughly to \code{random.rand()} from NumPy (\href{https://numpy.org/doc/stable/reference/random/generated/numpy.random.rand.html}{see}).

This function corresponds to \code{random.normal()} from NumPy (\href{https://numpy.org/doc/stable/reference/random/generated/numpy.random.normal.html}{see}).
}
