% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.r
\name{rearrange}
\alias{rearrange}
\title{Array rearrangement}
\usage{
rearrange(a, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{axis}{The axis or axes along \code{a} will be read into the new shape. The default \code{NULL} indicates the reverse order of all dimensions.}
}
\value{
The array \code{a} with swapped dimensions.
}
\description{
Rearrange an array.
}
\details{
Rearrangement of an array is equal to permutation its dimensions (axes). The given \code{axis} are the dimensions along the data of \code{a} are read.
The remaining axes span the dimension space where the data are read into, including \code{axis} at the last position of the entire dimension space (shape).
}
\seealso{
\code{\link{transpose}}.
}
