% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{rescale.array}
\alias{rescale.array}
\title{Array searching}
\usage{
rescale.array(a, ..., FUN)
}
\arguments{
\item{a}{An array.}

\item{...}{Indexing instructions with or without letters in form of \code{name = value} pairs. The names of the arguments specify the axis and the values its positions.}

\item{FUN}{Rescale function.}
}
\value{
An array with rescaled values.
}
\description{
Rescale elements of an array.
}
\examples{
a <- marray(seq(24), dim = c(4, 3, 2))
rescale.array(a, i = c(1, 3), j = 2, FUN = \(x) { x * 3 })

}
