% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{reshape.array}
\alias{reshape.array}
\title{Reshape an array}
\usage{
reshape.array(a, dim = NULL, order = c("C", "F"))
}
\arguments{
\item{a}{An array.}

\item{dim}{A integerish vector of new dimensions to be set on the array.}

\item{order}{The order in which elements of \code{a} should be read during rearrangement.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The (redimensioned) array \code{a}.
}
\description{
Reshape an array
}
\details{
This function corresponds to \code{reshape()} from NumPy.
}
\seealso{
\code{\link[reticulate]{array_reshape}}.
}
