% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{reshape_broadcast}
\alias{reshape_broadcast}
\title{Reshape arrays on basis of broadcasting}
\usage{
reshape_broadcast(..., axis = NULL)
}
\arguments{
\item{...}{Any number of objects.}

\item{axis}{The axis along operations like binding are planned.}
}
\value{
An array, or list of arrays, each with a dimension necessary for planned operations along axis.
}
\description{
Reshape arrays on basis of broadcasting
}
\seealso{
\code{\link{broadcastDIM}}.
}
