% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.r
\name{roll}
\alias{roll}
\title{Array circulation}
\usage{
roll(a, shift, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{shift}{The number of places by which elements are shifted. If a tuple, then axis must be a tuple of the same size,
and each of the given axes is shifted by the corresponding number. If a scalar while axis is a tuple of ints, then the same
value is used for all given axes. A positive number indicates a shift to the right, a negative number to the left.}

\item{axis}{The axis or axes along \code{a} which elements are shifted.
By default, the array is flattened before shifting, after which the original shape is restored.}
}
\value{
An array with the same shape as \code{a} but with shifted elements along given axis.
}
\description{
Roll array elements along a given axis.
}
\details{
This function corresponds to \code{roll()} from NumPy.
Shift array circularly, that is, elements that roll beyond the last position are re-introduced at the first.
}
\examples{
a <- marray(0:9)
# roll forward
roll(a, shift = 2)
# roll backward
roll(a, shift = -2)
a <- marray(a, dim = c(2, 5))
roll(a, shift = 1)
roll(a, shift = -1)
roll(a, shift = 1, axis = 1)
roll(a, shift = -1, axis = 1)

}
