% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.r
\name{row_stack}
\alias{row_stack}
\title{Array stack}
\usage{
row_stack(..., order = c("C", "F"))
}
\arguments{
\item{...}{Any numbers of objects they are coerced to 1-D arrays. The objects can be packed into a \code{\link{list}}. The length of these objects must be equal.}

\item{order}{The order in which elements of the objects should be read during coercing to arrays.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An 2-D array formed by stacking all input arrays.
}
\description{
Stack 1-D arrays as rows into a 2-D array.
}
\details{
This function corresponds to \code{row_stack()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.row_stack.html}{see}).
}
\examples{
a <- 1:3
b <- 4:6
row_stack(a, b)
}
