% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayio.r
\name{save.array}
\alias{save.array}
\alias{load.array}
\title{Array Serialization}
\usage{
save.array(..., file = "", fileext = ".rds")

load.array(file = "", fileext = ".rds")
}
\arguments{
\item{...}{Any number of arrays. The arrays can be packed into a \code{\link{list}} to make code more pipe-friendly.}

\item{file}{The names of the files or a folder the arrays are stored into or restored from.}

\item{fileext}{The extension of the stored array files, default is \code{.rds}.}
}
\value{
\code{NULL} invisibly.

An array or a list of restored arrays.
}
\description{
Save and load arrays.
}
\details{
If \code{file} is a folder, the names of the arguments are used as physical file names to store the arrays.
  For restoring, if \code{file} is a folder all files with the extension \code{fileext} are restored as arrays and packed into a list.
}
\examples{
a <- marray(1:12, dim = c(4, 3))
b <- marray(1:12, dim = c(4, 3, 1))
c <- marray(1:12, dim = c(4, 3, 1, 1))
save.array(a, b, c, file = getwd())

}
\seealso{
\code{\link{saveRDS}}.

\code{\link{readRDS}}.
}
