% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,MArray-method}
\alias{show,MArray-method}
\title{Show method for an MArray object}
\usage{
\S4method{show}{MArray}(object)
}
\arguments{
\item{object}{An object of class MArray}
}
\value{
Text describing the dimensions of an MArray object
}
\description{
Show method for an MArray object
}
\examples{
toy <- MArray::MakeMArrayDataSet()
toy_ma <-MArray(marray = toy$marray, fmeta = toy$fmeta, pmeta = toy$pmeta)
show(toy_ma)
}
