% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.r
\name{slice}
\alias{slice}
\alias{slice<-}
\title{Array slicing}
\usage{
slice(a, ..., drop = FALSE)

\code{slice(a, ...) <- value}.
}
\arguments{
\item{a}{A vector, matrix, or array.}

\item{...}{Indexing instructions with or without letters in form of \code{name = value} pairs. The names of the arguments specify the axis and the values its positions.}

\item{drop}{For matrices and arrays. If \code{TRUE} the result is coerced to the lowest possible dimension. This only works for extracting elements, not for the replacement. See \code{\link[base]{drop}} for further details.}

\item{value}{Any values to assign to the slice of \code{a}.}
}
\value{
An extracted part of \code{a}.
}
\description{
Read and write slices of an array using enhanced indexing operations.
}
\details{
\code{slice} is an alternative way to handle indexing array objects, usually done with \code{\link[base]{[}}. The dimensions can be indexed by names,
i for the first dimension, j for the second and so on. The assigned values are the positions on the corresponding dimension. The indexing expressions are the same as for \code{\link[base]{[}}.
}
\examples{
a <- marray(1:48, dim = c(4, 3, 2, 2))
slice(a) # complete four-dimensional array
slice(a, l = 2) # the values of the array of the second element of the last dimension (4th dimension)
slice(a, i = 1, j = 3) # the values of the array of the first element of the first dimension (1st row) and the third element of the second dimension (3rd column) across all bunches of the remaining dimensions 3 and 4.

a <- marray(1:24, dim = c(4, 3, 2), order = "F")
slice(a, i = 1L) <- 0L # write 0 to the first dimension (row) across all remaining dimensions
slice(a, i = 1L) <- 100:102 # write 100-102 to the first dimension (row) across all remaining dimensions
slice(a, i = 1L) <- 100:105 # write 100-105 to the first dimension (row) across all remaining dimensions
slice(a, i = 1L) <- matrix(100:105, nrow = 2L) # equal to prior, nrow can be 1, 2, 3, or 6
}
\references{
Indexing by letters is inspired by \url{https://github.com/cran/arrayhelpers}.
}
