% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.r
\name{slide}
\alias{slide}
\title{Array sliding}
\usage{
slide(a, size = 1L, stride = 1L, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{size}{An integerish vector indicating the size of the window sliding over \code{a}. A single value will be used for every \code{axis}.}

\item{stride}{An integerish vector specifying the amount of the window shift per each single move. A single value will be used for every \code{axis}.}

\item{axis}{The axis or axes along which to slide over \code{a}, default are all axes within the dimension space (shape). The not-considered axes are used without any limitation.}
}
\value{
A list of windowed sub-arrays.
}
\description{
Slides over an array with a window of given size and given stride.
}
\examples{
a <- marray(sample(7 * 4 * 2), dim = c(7, 4, 2))
arys <- slide(a) # sliding along each axes with a size of 1 and a stride of 1
arys <- slide(a, size = c(2, 1), axis = c(1, 2)) # sliding along first and second axes with sizes of 2 and 1 and a stride of 1 for both axes

}
