% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[,MArray-method}
\alias{[,MArray-method}
\title{Subsetting MArray objects}
\usage{
\S4method{[}{MArray}(x, i, j, drop = "missing")
}
\arguments{
\item{x}{An object of class "MArray"}

\item{i}{A row index value}

\item{j}{A column index value}

\item{drop}{Set to 'missing' to ignore irrelevant arguments}
}
\value{
A subset of an original "MArray" object
}
\description{
Subsetting MArray objects
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
sub_ma_obj <- ma_obj[3,4]
show(sub_ma_obj)
}
