% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.r
\name{swapaxes}
\alias{swapaxes}
\title{Array axis swapping}
\usage{
swapaxes(a, axis1, axis2)
}
\arguments{
\item{a}{An array.}

\item{axis1}{First axis.}

\item{axis2}{Second axis.}
}
\value{
The array \code{a} with swapped dimensions.
}
\description{
Interchange two axes of an array.
}
\details{
This function corresponds to \code{swapaxes()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.swapaxes.html}{see}).
}
\seealso{
\code{\link{transpose}}.
}
