% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.r
\name{take}
\alias{take}
\title{Array slicing}
\usage{
take(a, indices, axis = NULL)
}
\arguments{
\item{a}{An array.}

\item{indices}{The indices of the values to extract. Allows scalars as well as an array.}

\item{axis}{The axis over which to select values. By default, the flattened input array is used.}
}
\value{
An extracted part of \code{a}.
}
\description{
Take elements from an array along an axis.
}
\details{
This function corresponds to \code{take()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.take.html}{see}).
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
take(a, indices = c(11, 2, 23))
take(a, indices = marray(c(11, 2, 23), dim = c(1, 3)))
take(a, indices = marray(c(1, 3), dim = c(1, 2)), axis = 1)

}
