% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.r
\name{tile}
\alias{tile}
\title{Array repetition}
\usage{
tile(a, reps)
}
\arguments{
\item{a}{The input array.}

\item{reps}{Number of repetitions of \code{a} along each axis. If \code{reps} has length \code{d}, the result will have dimension of \code{max(d, ndim(a))}.}
}
\value{
The tiled output array.
}
\description{
Construct an array by repeating \code{a} the number of times given by \code{reps}.
}
\details{
This function corresponds to \code{tile()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.tile.html}{see}).
}
\examples{
a <- marray(c(0, 1, 2))
tile(a, 3)
tile(a, c(2, 2))
tile(a, c(2, 1, 2))

a <- marray(1:4, dim = c(2, 2))
tile(a, 2)
tile(a, c(2, 1))

}
\seealso{
\code{\link{marepeat}}.
}
