% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.r
\name{t.array}
\alias{t.array}
\alias{transpose}
\title{Array transposition}
\usage{
\method{t}{array}(a)

transpose(a, perm = NULL)
}
\arguments{
\item{a}{An array.}

\item{perm}{The permutation vector of the dimensions. The default \code{NULL} indicates to reverse the order of all dimensions.}
}
\value{
The array \code{a} with reversed or swapped dimensions.
}
\description{
Transpose an array.
}
\seealso{
\code{\link{t}}, \code{\link{rearrange}}.
}
