% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{trim_zeros}
\alias{trim_zeros}
\title{Array manipulation}
\usage{
trim_zeros(filt, trim = c("fb", "f", "b"))
}
\arguments{
\item{filt}{1D input array.}

\item{trim}{A string with ‘f’ representing trim from front and ‘b’ to trim from back. Default is ‘fb’, trim zeros from both front and back of the array.}
}
\value{
A 1D array of trimming the input.
}
\description{
Trim the leading and/or trailing zeros from a 1D array.
}
\details{
This function corresponds to \code{trim_zeros()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.trim_zeros.html}{see}).\cr
}
\examples{
a <- marray(c(0, 0, 0, 1, 2, 3, 0, 2, 1, 0))
trim_zeros(a)
trim_zeros(a, 'b')

}
