% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{vander}
\alias{vander}
\title{Array creation}
\usage{
vander(data, n)
}
\arguments{
\item{data}{The data to be reshaped to a Vandermonde matrix.}

\item{n}{The number of columns of the resulting matrix.}
}
\value{
A Vandermonde matrix.
}
\description{
Create Vandermonde matrix.
}
\details{
This function corresponds to \code{vander()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.vander.html}{see}).
}
