% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{where}
\alias{where}
\title{Array searching}
\usage{
where(a, condition, true, false)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to search for.}

\item{true}{Replacement value or function if conditions are met.}

\item{false}{Replacement value or function if conditions aren't met.}
}
\value{
An array of same shape as \code{a} with elements either from \code{true} if conditions are met or from \code{false} elsewhere.
}
\description{
Manipulate array elements depending on conditions.
}
\details{
This function corresponds to \code{where()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.where.html}{see}).
  If \code{true} or \code{false} are \code{NULL} the original values remain. For both arguments also functions can be specified.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
where(a, (a >= 11) & (a < 20), true = 23, false = 67)
where(a, a < 11, true = function(x) {x * 100}, false = 0)
where(a, a < 11, true = \(x) x * 100, false = 0) # lamdba-like anonymous function syntax \(x) from R 4.1

}
