% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{DIM}
\alias{DIM}
\alias{shape}
\title{Array shape}
\usage{
DIM(x)

shape(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The number of dimensions or in cases of an atomic object the length.
}
\description{
Retrieve dimensions of an object or its length.
}
\details{
The function \code{shape()} is a wrapper function for \code{DIM()}.
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
