% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.r
\name{embedseries}
\alias{embedseries}
\alias{embedseries.default}
\alias{embedseries.matrix}
\alias{embedseries.data.frame}
\title{Series embedding}
\usage{
embedseries(data, ...)

\method{embedseries}{default}(data, length = 1L, flip = TRUE, order = c("C", "F"))

\method{embedseries}{matrix}(data, length = 1L, flip = TRUE)

\method{embedseries}{data.frame}(data, length = 1L, flip = TRUE)
}
\arguments{
\item{data}{The data to be embedded into a shifted series.}

\item{length}{The length of a series.}

\item{flip}{A logical value indicating whether the embedded series should be reversed.}

\item{order}{The order in which elements of \code{data} should be read during flattening.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
An ongoing shifted series of \code{data}.
}
\description{
Series embedding
}
\examples{
x <- seq_len(1e+6)
df <- data.frame(a = seq_len(1e+6), b = -seq_len(1e+6))
a <- embedseries(x, length = 11L)
head(a)
a <- embedseries(df, length = 23L)
head(a)
}
