% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dims.r
\name{expand_dims}
\alias{expand_dims}
\title{Expand the shape of an array}
\usage{
expand_dims(a, axis = -1L)
}
\arguments{
\item{a}{An array.}

\item{axis}{Index position of the new axis in the expanded array. Negative numbers count from the back.}
}
\value{
The expanded array \code{a} with new shape.
}
\description{
Insert a new axis that will appear at the axis position in the expanded array shape.
}
\details{
This function corresponds to \code{expand_dims()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.expand_dims.html}{see}).
}
\references{
Implementation credits go to \url{https://github.com/t-kalinowski/listarrays}.
}
