% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{eye}
\alias{eye}
\title{Array creation}
\usage{
eye(n, m = NULL)
}
\arguments{
\item{n}{The number of rows.}

\item{m}{The number of columns, default \code{NULL}.}
}
\value{
An identity matrix with n rows and n or rather m columns.
}
\description{
Create 2D identity matrix.
}
\details{
This function corresponds to \code{eye()} from NumPy.
}
