% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fmeta}
\alias{fmeta}
\title{Accessor function for the 'fmeta' slot of an MArray object}
\usage{
fmeta(object)
}
\arguments{
\item{object}{An object of class MArray}
}
\description{
Accessor function for the 'fmeta' slot of an MArray object
}
