% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marray.r
\name{foreach.array}
\alias{foreach.array}
\alias{as.marray_int}
\alias{as.marray_dbl}
\alias{as.marray_raw}
\alias{as.marray_cpx}
\alias{as.marray_chr}
\alias{as.marray_lgl}
\alias{as.marray_norm}
\title{Array function application}
\usage{
foreach.array(a, FUN, ...)

as.marray_int(a)

as.marray_dbl(a)

as.marray_raw(a)

as.marray_cpx(a)

as.marray_chr(a)

as.marray_lgl(a)

as.marray_norm(a, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{a}{An array.}

\item{FUN}{The function to be applied on each element of \code{a}.}

\item{...}{Optional arguments passed to \code{FUN}.}

\item{mean}{Vector of means.}

\item{sd}{Vector of standard deviations.}

\item{log}{A logical value indicating whether the probabilities are given as \code{\link{log}}.}
}
\value{
The array \code{a} with elements processed by \code{FUN}.
}
\description{
Array function application
}
\details{
The funcctions \code{as.marray_int()}, \code{as.marray_dbl()}, \code{as.marray_raw()}, \code{as.marray_cpx()}, \code{as.marray_chr()} and \code{as.marray_lgl()}
transform the elements of an array into its corresponding types \code{\link{integer}}, \code{\link{double}}, \code{\link{raw}}, \code{\link{complex}}, \code{\link{character}} and \code{\link{logical}}.
}
\examples{
a <- marray(seq(24), dim = c(4, 3, 2))
foreach.array(a, FUN = function(x) 3 * x)
as.marray_int(a)
as.marray_dbl(a)
as.marray_raw(a)
as.marray_chr(a)
as.marray_lgl(a)
as.marray_norm(a)

}
