% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{full}
\alias{full}
\title{Array creation}
\usage{
full(dim = NULL, fill_value = NA, dimnames = NULL, order = c("C", "F"))
}
\arguments{
\item{dim}{Shape of the new array.}

\item{fill_value}{Value to fill the array.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of the length given by \code{dim} for that dimension.}

\item{order}{Whether to store multidimensional array in C- or Fortran-contiguous (row- or column-wise) order.}
}
\value{
An array of \code{fill_value} with the given shape.
}
\description{
Create array filled with value.
}
\details{
This function corresponds to \code{full()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.full.html}{see}).
}
