% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MArray-class}
\alias{MArray-class}
\alias{MArray}
\title{Constructor for making an instance of an 'MArray' object}
\arguments{
\item{marray}{matrix}

\item{pmeta}{dataframe}

\item{fmeta}{dataframe}
}
\value{
Object of class "MArray"
}
\description{
Constructor for making an instance of an 'MArray' object
}
\examples{
ma <- MakeMArrayDataSet(n_samples = 16L, n_features = 12L, with_seed = 8675)
ma_obj <- MArray(marray = ma$marray, fmeta = ma$fmeta, pmeta = ma$pmeta)
show(ma_obj)
}
