% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{nsize}
\alias{nsize}
\title{Array size}
\usage{
nsize(x)
}
\arguments{
\item{x}{A multidimensional data structure like array, matrix or data.frame.}
}
\value{
Number of elements.
}
\description{
Number of elements.
}
\details{
This function corresponds to \code{ndarray.size} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.ndarray.size.html}{see}).
}
