% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{pmeta}
\alias{pmeta}
\title{Accessor function for the 'pmeta' slot of an MArray object}
\usage{
pmeta(object)
}
\arguments{
\item{object}{An object of class MArray}
}
\description{
Accessor function for the 'pmeta' slot of an MArray object
}
