% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.r
\name{resize.array}
\alias{resize.array}
\title{Resize an array}
\usage{
resize.array(
  a,
  dim,
  fill = c("copy", "zero", "na", "approx"),
  order = c("C", "F")
)
}
\arguments{
\item{a}{An array.}

\item{dim}{An integerish vector of new shape (dimension space) to be set on the array.}

\item{fill}{The type of fill method. If the new array is larger than the original array, then the new array is filled with:
\itemize{
\item \code{copy}: repeated copies of \code{a}
\item \code{zero}: zeros
\item \code{na}: NA
\item \code{approx}: approximated values
}}

\item{order}{The order in which elements of \code{a} should be read during rearrangement.
By default, the order is equivalent to the \code{C}-style ordering and means elements should be read in row-major order.
In opposite, the \code{Fortran}-style ordering means elements should be read in column-major order.}
}
\value{
The new array \code{a} with given shape.
}
\description{
Resize an array
}
\details{
This function corresponds partially to \code{resize()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.resize.html}{see}). The function from NumPy only uses the copy type.
}
