% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.r
\name{rot90}
\alias{rot90}
\title{Array rotation}
\usage{
rot90(a, k = 1L, axes = c(1L, 2L))
}
\arguments{
\item{a}{An array.}

\item{k}{Number of times the array is rotated by 90 degree. Positive numbers represent clockwise rotation, negative numbers counterclockwise rotation.}

\item{axes}{The array is rotated in the plane defined by the axes. Axes must be different.}
}
\value{
A rotated view of \code{a}.
}
\description{
Rotate an array by 90 degrees in the plane specified by axes.
}
\details{
This function corresponds to \code{rot90()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.rot90.html}{see}).
}
