% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.r
\name{where}
\alias{where}
\title{Array searching}
\usage{
where(a, condition, true, false)
}
\arguments{
\item{a}{An array.}

\item{condition}{Logical expression indicating elements to search for.}

\item{true}{Replacement value or function if conditions are met.}

\item{false}{Replacement value or function if conditions aren't met.}
}
\value{
An array of same shape as \code{a} with elements either from \code{true} if conditions are met or from \code{false} elsewhere.
}
\description{
Manipulate array elements depending on conditions.
}
\details{
This function corresponds to \code{where()} from NumPy (\href{https://numpy.org/doc/stable/reference/generated/numpy.where.html}{see}). \cr

The result is an array with the same shape as \code{a}, where the elements in \code{a} that satisfy or do not satisfy the given condition are processed in a specific way and placed at the same position as in \code{a} in the result.
Elements that satisfy the condition are processed according to the specification for \code{true}, elements that do not satisfy the condition are processed according to the specification for \code{false}.
If \code{true} or \code{false} are \code{NULL} the original values from \code{a} remain. For both arguments concrete values as well as functions can be specified.
}
\examples{
a <- marray(1:24, dim = c(4, 3, 2))
where(a, (a >= 11) & (a < 20), true = 23, false = 67)
where(a, a < 11, true = function(x) {x * 100}, false = 0)
where(a, a < 11, true = \(x) x * 100, false = 0) # lamdba-like anonymous function syntax \(x) from R 4.1

}
