% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.r
\name{zeros}
\alias{zeros}
\title{Array creation}
\usage{
zeros(dim = NULL, dimnames = NULL)
}
\arguments{
\item{dim}{Shape of the new array.}

\item{dimnames}{Either \code{NULL} or the names of the dimensions. This must be a list with one component for each dimension, either \code{NULL} or a character vector of the length given by \code{dim} for that dimension.}
}
\value{
An array of zeros with the given shape.
}
\description{
Create array filled with zeros.
}
\details{
This function corresponds to \code{zeros()} from NumPy.
}
